% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{addGmatToSnap}
\alias{addGmatToSnap}
\title{Add cell-by-gene matrix}
\usage{
addGmatToSnap(obj, num.cores)
}
\arguments{
\item{obj}{A snap object to add cell-by-bin matrix.}

\item{num.cores}{Number of processors to use.}
}
\value{
Return a snap object
}
\description{
This function takes a snap object as input and add the cell-by-gene 
matrix to the existing snap object.
}
