% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{addPmatToSnap}
\alias{addPmatToSnap}
\title{Add cell-by-peak matrix}
\usage{
addPmatToSnap(obj, num.cores)
}
\arguments{
\item{obj}{A snap object.}

\item{num.cores}{Number of processors to use.}
}
\description{
This function takes a snap object as input and add the 
cell-by-peak matrix to the existing snap object.
}
