% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{barcodeInSnapFile}
\alias{barcodeInSnapFile}
\title{Check barcode existance in snap file}
\usage{
barcodeInSnapFile(barcode, file)
}
\arguments{
\item{barcode}{An array of selected barcodes.}

\item{file}{A snap format file.}
}
\value{
Return an array of logical variable indicates whether the 
barcode exists in snap file.
}
\description{
This function takes an array of barcodes and a snap-format file as input 
and check whether selected barcodes exist in the snap file.
}
