% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{boxPlotFeature}
\alias{boxPlotFeature}
\title{Feature Enrichment Boxplot}
\usage{
boxPlotFeature(obj, feature, outline, ylab, main, add.point, point.size,
  point.shape, point.alpha, pdf.file.name, pdf.height, pdf.width)
}
\arguments{
\item{obj}{A snap object.}

\item{feature}{Feature enrichment value for each cell.}

\item{outline}{If 'outline' is not true, the outliers are not drawn (as points whereas S+ uses lines).}

\item{ylab}{Name of ylab.}

\item{main}{Main title.}

\item{add.point}{If 'add.point' is true, scatter points will be added to the top of the boxplot.}

\item{point.size}{Point size [0.5].}

\item{point.shape}{Point shape type [19].}

\item{point.alpha}{Point transparancy level [0.9].}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{pdf.width}{the width of the graphics region in inches [7].}
}
\description{
Feature Enrichment Boxplot
}
\examples{
data(demo.sp);
boxPlotFeature(demo.sp, rnorm(nrow(demo.sp)));

}
