% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calGmat.R
\name{calCellGeneTable}
\alias{calCellGeneTable}
\title{Calculate Cell x Gene Table from a snap Object}
\usage{
calCellGeneTable(object, gene, ...)
}
\arguments{
\item{object}{A Snap object.}

\item{gene}{A GenomicRange object that contains the genomic intervals of interest.}

\item{norm}{A logical class that indicates whether to normalize the count matrix. If True, the count table will be normalized to FPKM (default TRUE).}

\item{ncores}{A numeric class that indicates how many processers to use for computation (default 1). The more used, the faster it runs.}
}
\value{
Returns a Snap object with the cell x gene matrix stored in object@gmat
}
\description{
This function takes a snap object and a GenomicRanges object that indicates the 
features or genes locations of interest and returns a cell x gene count table.
}
