% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selBinSize.R
\name{checkBinSize}
\alias{checkBinSize}
\title{checkBinSize}
\usage{
checkBinSize(obj1, ...)
}
\arguments{
\item{obj}{character. a snap object}
}
\description{
This function takes a snap object as input and check if the current bin size is ok
}
