% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDuplicate.R
\name{checkDuplicate}
\alias{checkDuplicate}
\title{Check & Remove Duplicates}
\usage{
checkDuplicate(x, ...)
}
\arguments{
\item{x}{the numeric matrix that provides data.}

\item{sd}{standard deviation for the noise.}
}
\description{
This function tables a data matrix as input and check weather duplicates exist. If so, add white noise to the duplicates
}
