% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkFilesSnapFiles}
\alias{checkFilesSnapFiles}
\title{Check if all the files are snap files.}
\usage{
checkFilesSnapFiles(snapFiles, do.par = FALSE, num.cores = 1)
}
\arguments{
\item{snapFiles}{list of string}

\item{do.par}{bool, if do in parallel, default is FALSE.}

\item{num.cores}{integer, cores to be used, default is 1}
}
\value{
bool, TRUE if all of them are snap files; FALSE if no
}
\description{
Check if all the files are snap files.
}
