% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkSnapFileSession}
\alias{checkSnapFileSession}
\title{Check if all snap files have a given session.}
\usage{
checkSnapFileSession(snapFiles, session, do.par = FALSE, num.cores = 1)
}
\arguments{
\item{snapFiles}{list of string}

\item{session}{character}

\item{do.par}{bool, if do in parallel, default is FALSE.}

\item{num.cores}{integer, cores to be used, default is 1}
}
\value{
bool, TRUE if all snap files have the session; FALSE if no.
}
\description{
Check if all snap files have a given session.
}
