% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkSnapFilesExist}
\alias{checkSnapFilesExist}
\title{Check if all the snap files exist.}
\usage{
checkSnapFilesExist(snapFiles)
}
\arguments{
\item{snapFiles}{list of string}
}
\value{
bool, TRUE if all of them exist; FALSE if not.
}
\description{
Check if all the snap files exist.
}
