% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\name{colCovs}
\alias{colCovs}
\title{colCovs for snap object.}
\usage{
colCovs(x, ...)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns the number of non-zero elements for each column of the count matrix
}
\examples{
barcode_num <- 100
feature_num <- 500
m <- matrix(sample(0:1,barcode_num * feature_num, replace=TRUE),barcode_num,feature_num)
x.sp <- newSnap(m)
feature_cov <- colCovs(x.sp)
}
