% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{colMeans}
\alias{colMeans}
\alias{colMeans,snap-method}
\title{colMeans for snap objects}
\usage{
\S4method{colMeans}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = FALSE)
}
\arguments{
\item{x}{snap; a snap object}
}
\description{
This function takes a snap object and returns the column means of its count matrix.
}
