% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\docType{methods}
\name{colSums}
\alias{colSums}
\alias{colSums,snap-method}
\title{colSums for snap objects}
\usage{
\S4method{colSums}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = TRUE)
}
\arguments{
\item{x}{A snap object}

\item{mat}{A charater object indicates what matrix slot to use c("bmat", "pmat", "gmat")}

\item{na.rm}{A logical variable indicates wether to remove NA in the matrix}
}
\description{
This function takes a snap object and returns the column sums of its count matrix.
}
\examples{
data(demo.sp);
a = colSums(demo.sp, mat="bmat");
b = colSums(demo.sp, mat="pmat");
d = colSums(demo.sp, mat="gmat");
}
