% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGmat.R
\name{createGmat}
\alias{createGmat}
\title{Create Cell-by-Gene Matrix}
\usage{
createGmat(obj, genes, ncell.chunk, do.par, num.cores)
}
\arguments{
\item{obj}{A snap object.}

\item{genes}{A GRanges object contains the gene annotation}

\item{ncell.chunk}{A numeric class that indicates the number of cells to calculate per processing core [20].}

\item{do.par}{A logical variable indicates if run this using multiple processors [TRUE].}

\item{num.cores}{Number of processers to use [1].}
}
\value{
Returns a snap objects contains cell-by-gene matrix
}
\description{
This function takes a snap object and gene list as input and creates cell-by-gene matrix
}
