% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{createSnap}
\alias{createSnap}
\title{Create a snap object from a snap file}
\usage{
createSnap(file, sample, description, do.par, num.cores)
}
\arguments{
\item{file}{Name of a snap-format file.}

\item{sample}{A short sample name (i.g. "MOS.rep1").}

\item{description}{Description of the experiment [NULL].}

\item{do.par}{A logical variable indicates if run this using multiple processors [FALSE].}

\item{num.cores}{Number of processers to use [1].}
}
\value{
A snap object
}
\description{
This function takes a snap-format file as input and create
a snap object.
}
\examples{
file.name = system.file("extdata", "demo.snap", package = "SnapATAC");
demo.sp = createSnap(file.name, sample="demo", do.par=FALSE);

}
