% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnap.R
\name{createSnapFromGmat}
\alias{createSnapFromGmat}
\title{Create a snap object from cell-by-gene matrix}
\usage{
createSnapFromGmat(mat, barcodes, geneNames)
}
\description{
Create a snap object from cell-by-gene matrix
}
