% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduct-class.R
\docType{class}
\name{dim.reduct-class}
\alias{dim.reduct-class}
\alias{dim.reduct}
\title{An S4 class to represent dimentionality reduction object.}
\description{
Class defines a dim.reduct object.
}
\section{Slots}{

\describe{
\item{\code{dmat}}{a matrix object that contains reduced dimentions}

\item{\code{sdev}}{variance for each principal conponents}

\item{\code{iter}}{iterations used for running dimentionality reduction}

\item{\code{method}}{a character object indicates the method used for dimentionality reduction}
}}

\examples{
data(demo.sp);
demo.sub.sp = demo.sp[1:5,]
}
