% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eig_decomp}
\alias{eig_decomp}
\title{Eig decomposition.}
\usage{
eig_decomp(M, n_eigs, sym = isSymmetric(M), method = "arpack")
}
\arguments{
\item{M}{Matrix}

\item{sym}{bool, is M symmetric, default is isSymmetric(M)}

\item{method}{character, "arpack" or "RSpectra" for decomposition, default is arpack}

\item{n_eighs}{integer}
}
\description{
Eig decomposition.
}
