% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{exportMetaData}
\alias{exportMetaData}
\title{Export barcode meta data}
\usage{
exportMetaData(obj, file, slot.names)
}
\arguments{
\item{obj}{A snap object}

\item{file}{Output file name}

\item{slot.names}{Name of slots to be exported c('barcode', 'tsne', 'umap', 'cluster', 'metaData')}
}
\description{
This function takes a snap object as input and export its barcode and corresponding attributes
}
