% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{extractReads}
\alias{extractReads}
\title{Extract Reads By Barcodes}
\usage{
extractReads(barcodes, files, do.par = TRUE, num.cores = 1)
}
\arguments{
\item{barcodes}{A vector contains the selected barcodes.}

\item{files}{A vector contains the snap file that barcodes belong to.}

\item{do.par}{A logic variable indicates weather to run this in parallel with multiple processors.}

\item{num.cores}{Number of processors to use.}
}
\value{
Returns A GenomicRanges object that contains the reads
}
\description{
This function takes a barcode list and snap file as input and quickly extract reads belonging to the given barcodes.
}
