% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells.R
\name{filterCells}
\alias{filterCells}
\title{Cell Filteration}
\usage{
filterCells(obj, ...)
}
\arguments{
\item{object}{snap object}

\item{subset.names}{Parameters to subset on. Eg, the name of a gene, PC1, a
column name in object@meta.data, etc. Any argument that can be retreived
using FetchData}

\item{low.thresholds}{Low cutoffs for the parameters (default is -Inf)}

\item{high.thresholds}{High cutoffs for the parameters (default is Inf)}
}
\value{
Returns a snap object containing only the relevant subset of cells
}
\description{
This function takes a snap object as input and filter nuclei based on given cutoff.
}
\examples{
head(x = FetchData(object = pbmc_small, vars.all = 'LTB'))
mos_filtered <- filterCells(
  object = mos,
  subset.names = 'fragment.num',
  low.thresholds = 1000
)
head(x = FetchData(object = pbmc_filtered, vars.all = 'LTB'))

}
