% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{findDAR.fisher}
\alias{findDAR.fisher}
\title{Identify Statistically Singficant Differential Accessible Regions Between 2 Clusters Using Fisher Exact Test}
\usage{
findDAR.fisher(object, input = c("cmat", "gmat"), idx = NULL, idy = NULL,
  psudo.count = TRUE, min.cov = 0.01, min.fc = 2, max.cov = 0.8,
  norm = TRUE, step = 1)
}
\arguments{
\item{object}{A Snap object}

\item{input}{A character object that indicates what attribute of snap object to use for analysis ("cmat" by default, or "gmat", "bmat")}

\item{idx}{A vector that contains index for experiment group of cells}

\item{idy}{A vector that contains index for control group of cells}

\item{psudo.count}{A logical variable that indicates whether to add psudo count}

\item{min.cov}{A numeric variable between 0 to 1 that indicates min coverage (0.01) for feature to be tested. Any features are covered less than min.cov of all cells are not going to be tested}

\item{min.fc}{A numeric variable that indicates min fold change (2) for feature to be tested. Any features with fold change less than min.fc will be not be tested}

\item{max.cov}{A numeric variable between 0 to 1 that indicates max coverage (0.6) for feature to be tested. Any features are covered more than max.cov of all cells are not going to be tested}

\item{norm}{A logical variable that indicates whether to adjust for sequencing depth between two clusters}

\item{step}{A numeric variable used when norm is TRUE [1]}
}
\value{
Returns a data.frame table that contains differential analysis info for feature:
id - index for tested features (because not all features have been tested)
p1 - feature coverage for treatment group
p2 - feature coverage for control group (adjusted for sequencing depth if norm is TRUE)
p - average coverage among both control and treatment group (p1+p2)/2
log2FC - log2 fold change between treatment and control
pval - Fisher exact test P-value
}
\description{
This function takes a snap object and index (idx and idy) for control and treatment cells as input and identify differential accessibile regions (DARs) features between 2 groups or clusters of cells.
}
