% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDARs.R
\name{findDAR}
\alias{findDAR}
\title{Identifiy Statistically Singficant Differential Accessible Regions (DARs)}
\usage{
findDAR(object, input = c("cmat", "gmat"), cluster = NULL, rep = NULL,
  treat = NULL, ctrl = NULL, down.sample = TRUE, dispersion = 0.2,
  min.row.sum = 10, max.row.sum = 500, rand.seed = 10, ...)
}
\arguments{
\item{object}{A Snap object}

\item{cluster}{A vector that indicates the cluster label for each cell}

\item{rep}{A vector that indicates the replicate label for each cell}

\item{treat}{A character or numeric object indicates which cluster or clusters are used as treatment group}

\item{ctrl}{A character or numeric object indicates which cluster or clusters are used as control group}

\item{down.sample}{A logic object indicates whether or not to perform down sampling of bigger cluster}

\item{dispersion}{A numberic object indicates the common dispersion for EXACT test}

\item{min.row.sum}{A numberic object that filters any feature with coverage less than min.row.sum}

\item{max.row.sum}{A numberic object that filters any feature with coverage more than min.row.sum}

\item{rand.seed}{A numeric object that sets the random seeds}

\item{method}{A character object indicates what test to use (LRT, EXACT, QLF)}
}
\value{
Returns a data.frame table that contains differential analysis info for each feature:
id - index for original region
cpmT - cpmC - averaged CPM for treatment and control
logFC - log2 fold change between treatment and control
logCPM - log2 average CPM between treatment and control
LR - Likilihood Ratio
PValue - Chi-square P-value
fdr - Adjusted P-value
}
\description{
This function takes a snap object and a cluster vector and replicate indicator as input
and identify the differential accessibile regions (DARs) features between 2 groups of cells.
}
