% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getBarcode}
\alias{getBarcode}
\alias{getBarcode,snap-method}
\alias{getMetaData,snap-method}
\alias{getBarcode,snap-method}
\title{getBarcode for snap object.}
\usage{
getBarcode(obj)

\S4method{getBarcode}{snap}(obj)

\S4method{getMetaData}{snap}(obj)
}
\arguments{
\item{obj}{A snap object}
}
\description{
This function takes a snap object and returns the barcode information
}
\examples{
fname <- system.file("extdata", "demo_snap", package = "SNAPATAC")
load(fname)
barcodes <- getBarcode(schep.sp)
}
