% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getBinMatrix}
\alias{getBinMatrix}
\alias{getBinMatrix,snap-method}
\title{getBinMatrix for snap object.}
\usage{
getBinMatrix(obj)

\S4method{getBinMatrix}{snap}(obj)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns the cellxbin matrix
}
