% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getFeature}
\alias{getFeature}
\alias{getFeature,snap-method}
\title{getFeature for snap object.}
\usage{
getFeature(obj)

\S4method{getFeature}{snap}(obj)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns the features
}
\examples{
barcode_num <- 100
feature_num <- 500
m <- matrix(sample(0:1,barcode_num * feature_num, replace=TRUE),barcode_num,feature_num)
x.sp <- newSnap(m)
features <- getFeature(x.sp)
}
