% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getGraph}
\alias{getGraph}
\alias{getGraph,snap-method}
\alias{nrow}
\alias{for}
\alias{snap}
\alias{object.}
\alias{This}
\alias{function}
\alias{takes}
\alias{a}
\alias{object}
\alias{and}
\alias{returns}
\alias{number}
\alias{of}
\alias{barcodes}
\title{getGraph for snap object.}
\usage{
\S4method{nrow}{snap}(x)
}
\arguments{
\item{x}{snap; a snap object}

\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns its consensus KNN-jaccard graph
}
