% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getJaccardMatrix}
\alias{getJaccardMatrix}
\alias{getJaccardMatrix,snap-method}
\title{getJaccardMatrix for snap object.}
\usage{
getJaccardMatrix(obj)

\S4method{getJaccardMatrix}{snap}(obj)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns the jaccard matrix
}
