% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\name{getMetaData}
\alias{getMetaData}
\title{getMetaData for snap object.}
\usage{
getMetaData(obj)
}
\arguments{
\item{obj}{A snap object}
}
\description{
This function takes a snap object and returns the metaData information
}
\examples{
fname <- system.file("extdata", "demo_snap", package = "SNAPATAC")
load(fname)
barcodes <- getMetaData(schep.sp)
}
