% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getPCA}
\alias{getPCA}
\alias{getPCA,snap-method}
\title{getPCA for snap object.}
\usage{
getPCA(obj)

\S4method{getPCA}{snap}(obj)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns the pca cell.embedding
}
