% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getPeakMatrix}
\alias{getPeakMatrix}
\alias{getPeakMatrix,snap-method}
\title{getPeakMatrix for snap object.}
\usage{
getPeakMatrix(obj)

\S4method{getPeakMatrix}{snap}(obj)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns the cellxpeak matrix
}
