% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getTSNE}
\alias{getTSNE}
\alias{getTSNE,snap-method}
\title{getTSNE for snap object.}
\usage{
getTSNE(obj)

\S4method{getTSNE}{snap}(obj)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns its tsne coordinate
}
