% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{getUMAP}
\alias{getUMAP}
\alias{getUMAP,snap-method}
\title{getUMAP for snap object.}
\usage{
getUMAP(obj)

\S4method{getUMAP}{snap}(obj)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns its tsne coordinate
}
