% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{is.snap}
\alias{is.snap}
\alias{is.snap,snap-method}
\title{isSnap for snap object.}
\usage{
is.snap(x)

\S4method{is.snap}{snap}(x)
}
\arguments{
\item{x}{any R object.}
}
\description{
Functions to check if an object is a snap object
}
