% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSnapFile.R
\name{isSnapFile}
\alias{isSnapFile}
\alias{isSnapFile.default}
\title{Snap}
\usage{
isSnapFile(file)

\method{isSnapFile}{default}(file)
}
\arguments{
\item{file}{character; A character indicates the file name}
}
\value{
Weather a given file is a snap-formatted file
}
\description{
This function takes a file name as input and check if the file
is a snap-formated file
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default Interface
}}

\examples{
fname <- system.file("extdata", "Schep.snap", package = "nucleus")
isSnapFile(fname)

}
