% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard-class.R
\docType{class}
\name{jaccard-class}
\alias{jaccard-class}
\title{An S4 class jaccard to represent a jaccard object.}
\description{
Class jaccard defines a jaccard object.
}
\section{Slots}{

\describe{
\item{\code{jmat}}{a matrix object that contains the jaccard index similarity matrix}

\item{\code{p1}}{an array of numeric values indicates the coverage for rows}

\item{\code{p2}}{an array of numeric values indicates the coverage for columns}

\item{\code{norm}}{a logical variable indicates whether jaccard index matrix has been normalized}

\item{\code{method}}{normalization method used to normalize jaccard index matrix}
}}

