% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgraph-class.R
\docType{class}
\name{kgraph-class}
\alias{kgraph-class}
\title{An S4 class jaccard to represent a knn graph object.}
\description{
Class kgraph defines a kgraph object.
}
\section{Slots}{

\describe{
\item{\code{mat}}{a parase matrix object that contains the adjacent matrix of knn graph}

\item{\code{k}}{k used for constructing knn graph}

\item{\code{snn}}{knn is converted to snn graph if True}

\item{\code{snn.prune}}{snn prunning}
}}

