% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBinary.R
\name{makeBinary}
\alias{makeBinary}
\title{Convert the count matrix to binary matrix}
\usage{
makeBinary(object, ...)
}
\arguments{
\item{object}{A snap object.}

\item{outlier.filter}{A numeric class within the range between 0 and 1. The non-zero elememnts in the count matrix are identified annd filtered as outliers (default 1e-3)}
}
\value{
Returns a Snap object with the binary matrix stored in object@bmat
}
\description{
This function takes a snap object as input, then convert count matrix to a binary matrix.
}
\details{
It has been observed that some entries in the count matrix have exceedingly high value, which we suspect 
are due to alingnment or other type of errors. This function first identifies the top outlier.filter (%) 
non-zero values in the count matrix as cutoff and then remove any values above this cutoff as outliers.
}
\examples{
schep.sp
schep.sp = cmat2bmat(schep.sp, outlier.filter=1e-3);

}
