% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mos}
\alias{mos}
\title{Single Nucleus ATAC-seq for MOs 2K}
\format{An object of class \code{"snap"} with 2000 cells:
\describe{
  \item{metData}{meta data of each cell}
  \item{bmat}{cell x bin matrix}
  \item{pmat}{cell x peak matrix}
  \item{gmat}{cell x gene matrix}
  \item{jmat}{jaccard index matrix object}
  \item{graph}{knn graph}
  \item{cluster}{cluster result}
  \item{tsne}{tsne cooridnates}
  \item{umap}{umap cooridnates}
}}
\usage{
data(mos)
}
\description{
Data from single nucleus ATAC-seq experiment on secondary motor cortex
in adult mouse brain. The data is down-sampled to 2,000 cells from the 
original dataset.
}
\examples{
data(mos)
}
\keyword{datasets}
