% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{ncol}
\alias{ncol}
\alias{ncol,snap-method}
\title{ncol for snap object.}
\usage{
\S4method{ncol}{snap}(x)
}
\arguments{
\item{x}{snap; a snap object}
}
\description{
This function takes a snap object and returns number of features
}
