% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\name{newSnap}
\alias{newSnap}
\title{Create a snap object}
\usage{
newSnap(...)
}
\arguments{
\item{cmat}{Matrix; A count matrix with barcodes as rows and genomic features as columns}

\item{barcode}{character; A vector object contains barcodes for the rows of mat}

\item{feature}{GRanges; A GRanges object contains genomic bins for the column of mat}
}
\value{
A snap object
}
\description{
This function takes a matrix of single-nucleus count matrix 
(cell in rows and genomic features as columns) and a vector 
of barcodes and a GRanges of genomic features (genes or bins) 
to create a snap object
}
\examples{
barcode_num = 100
feature_num = 500
m <- matrix(sample(0:1,barcode_num * feature_num, replace=TRUE),barcode_num,feature_num)
newSnap(m)

}
