% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normJaccard.R
\name{normJaccard.default}
\alias{normJaccard.default}
\title{Normalize Jaccard Index Matrix}
\usage{
normJaccard.default(object, k = 15, outlier.filter = 0.001, eps = 1e-06,
  cell.pcore = 1000, ncore = 5, method = c("normOVE", "normOVN"),
  row.center = TRUE, row.scale = TRUE)
}
\arguments{
\item{object}{A snap object}

\item{k}{A numeric class - number of neibouring cells to use for OVN (used only if method="OVN") (default k = 15)}

\item{outlier.filter}{A numeric class that indicates the top outlier.filter of values as outliers to be capped (0.01)}

\item{method}{A character class that indicates the normalization method to be used. This must be one of "OVN", "OVE"}

\item{max.cell}{A numeric class that indicates the max number of cells to calculate per core}

\item{cores}{A numeric class that indicates the number of cores to use for calculation}

\item{row.norm}{a logical value indicating whether rows of the normalized jaccard index matrix should be centered to 0.}
}
\description{
This function takes a Snap object as input with jmat attributes and normalize 
for read depth effect.
}
\details{
In theory, the entry in the jaccard index calculated by calJaccard() should 
reflects the true similarity between two cells, however, that is not the case. 
We observed that a cell of higher coverage tends to have a higher similarity 
with another cell regardless whether these two cells are similar or not.
These biases, we termed as “coverage bias” also observed in other studies, 
can later result in misleading cell grouping. Therefore, it is cruicial to 
normalize the bias.

Here we propose three different normalization methods.

1. Observe Over Neighbours (OVN)
2. Observe Over Expected   (OVE)

1. Observe Over Neighbours (OVN)
For each pair of cells i and j, we first identify their neibours Ni and Nj based
on the coverage. Second, we calcualted the pair-wise jaccard index between Ni 
Nj as Eij. The average mean(Eij) is considered to be the expected jaccard index 
between cell i and j. The normalized jaccard index between i and j is Oij - mean(Eij).

2. Observe Over Expected (OVE)
Alternatively, one can calcualte the theoritical expected jaccard index between 
cell i and j if the "1"s are completely random. However, we found the theoretically
estimated expected jaccard index is usually lower than that estimated from the neibours.
This is also expected because the accessible sites are not completely random. For instance,
we previously found the promoters of housekeeping genes are more often occupied between
different cells. To solve this under-estimate problem for OVE, we found that OVE is highly
correlated with OVN (r>0.99), therefore, we can estimate a scale factor alpha to scale 
OVE to similar level with OVN. The scale factor is estimated from the data.
}
