% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDA.R, R/runTF_IDF.R
\name{normTFIDF}
\alias{normTFIDF}
\alias{normTFIDF}
\title{Frequency–inverse document frequency (IF-IDF)}
\usage{
normTFIDF(object, ...)

normTFIDF(object, ...)
}
\arguments{
\item{object}{A snap object}

\item{object}{A snap object}
}
\description{
This function takes a Snap object as input and perform IF-IDF against the binary matrix.

This function takes a Snap object as input and perform IF-IDF against the binary matrix.
}
\details{
Latent Semantic Indexing (LSI), a technique commonly used in the natural language processing 
that analyze the relationships between a set of documents, has been applied to scale the 
binary matrix and infer the association between cells. LSI first created a TF-ITD 
(frequency–inverse document frequency) matrix which scales the binary matrix by weighting 
each site accessible in an individual cell by the total number of sites accessible in that cell.

Latent Semantic Indexing (LSI), a technique commonly used in the natural language processing 
that analyze the relationships between a set of documents, has been applied to scale the 
binary matrix and infer the association between cells. LSI first created a TF-ITD 
(frequency–inverse document frequency) matrix which scales the binary matrix by weighting 
each site accessible in an individual cell by the total number of sites accessible in that cell.
}
