% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\docType{methods}
\name{nrow}
\alias{nrow}
\alias{nrow,snap-method}
\title{nrow for snap object.}
\usage{
\S4method{nrow}{snap}(x)
}
\arguments{
\item{x}{snap; a snap object}
}
\description{
This function takes a snap object and returns number of cells
}
\examples{
data(demo);
nrow(demo);

}
