% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotBarcode}
\alias{plotBarcode}
\title{Plot Barocde Quality Control Distribution}
\usage{
plotBarcode(obj, pdf.file.name, pdf.width, pdf.height, ...)
}
\arguments{
\item{obj}{a snap object}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.width}{the width of the graphics region in inches [7].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{...}{Arguments passed to hist method.}
}
\description{
Plot Barocde Quality Control Distribution
}
