% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotBinCoverage}
\alias{plotBinCoverage}
\title{Plot Bin Coverage Distribution}
\usage{
plotBinCoverage(obj, rm.zeros, pdf.file.name, pdf.width, pdf.height, ...)
}
\arguments{
\item{obj}{a snap object}

\item{rm.zeros}{Remove bins of zero coverage}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.width}{the width of the graphics region in inches [7].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{...}{Arguments passed to hist function.}
}
\description{
Plot Bin Coverage Distribution
}
