% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotDimReductElbow}
\alias{plotDimReductElbow}
\title{Elbow Plot for Dimentionality Reduction Result}
\usage{
plotDimReductElbow(obj, point.size, point.shape, point.color, point.alpha,
  pdf.file.name, pdf.height, pdf.width, ...)
}
\arguments{
\item{obj}{A snap object}

\item{point.size}{Point size [2].}

\item{point.shape}{Point shape type [19].}

\item{point.color}{Point color ["red"].}

\item{point.alpha}{Point transparancy level [1].}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{pdf.width}{the width of the graphics region in inches [7].}

\item{...}{Arguments passed to plot function.}
}
\description{
Elbow Plot for Dimentionality Reduction Result
}
\examples{
data(demo.sp);
demo.sp = makeBinary(demo.sp, mat="bmat");
demo.sp = runJaccard(
obj=demo.sp, 
tmp.folder=tempdir(), 
mat="bmat"
);
demo.sp = runNormJaccard(
obj=demo.sp, 
tmp.folder=tempdir()
);
demo.sp = runDimReduct(
obj=demo.sp, 
pc.num=10, 
input.mat="jmat"
);
plotDimReductElbow(demo.sp);

}
