% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{PlotFeatureSingle}
\alias{PlotFeatureSingle}
\title{PlotFeatureSingle}
\usage{
PlotFeatureSingle(obj, feature.value, method, point.size, point.shape,
  point.color, down.sample, pdf.file.name, pdf.width, pdf.height, ...)
}
\arguments{
\item{obj}{A snap object.}

\item{feature.value}{Feature enrichment value for each cell. Value will be normalized betweeen 0 and 1.}

\item{method}{Visulization method c("tsne", "umap").}

\item{point.size}{Point size [1].}

\item{point.shape}{Point shape type [19].}

\item{point.color}{Color of point ["red"].}

\item{down.sample}{Downsample the original cells to down.sample cells to ovoid large dataset [10,000].}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.width}{the width of the graphics region in inches [7].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{...}{Arguments passed to plot method.}
}
\description{
PlotFeatureSingle
}
\examples{
data(demo.sp);
x = rexp(nrow(demo.sp));
x = (x-min(x))/(max(x)-min(x));
PlotFeatureSingle(
	obj=demo.sp, 
	feature.value=x,
	method="tsne", 
	point.size=1, 
	point.shape=19, 
	point.color="red", 
	down.sample=10000, 
	pdf.file.name=NULL, 
	pdf.width=7, 
	pdf.height=7
)

}
