% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotGene}
\alias{plotGene}
\title{Project Gene-body Accessibility to T-SNE plot}
\usage{
plotGene(obj, gene.names, viz.method, point.size, point.color, point.shape,
  background.point, background.point.color, background.point.alpha,
  background.point.size, background.point.shape, low.value, high.value,
  down.sample, seed.use, plot.nrow, plot.ncol, pdf.file.name, pdf.height,
  pdf.width, ...)
}
\arguments{
\item{obj}{A snap object.}

\item{gene.names}{Name of genes to plot.}

\item{viz.method}{Project method c("tsne", "umap").}

\item{point.size}{Point size [0.5].}

\item{point.color}{Point color ["blue"].}

\item{point.shape}{Point shape type [19].}

\item{background.point}{If add points as background [TRUE].}

\item{background.point.color}{Color of background points ["grey"].}

\item{background.point.alpha}{Transparency level of background points [0.3].}

\item{background.point.size}{Size of background points [0.5].}

\item{background.point.shape}{Shape of background points [19].}

\item{low.value}{Feature value is standarded to 0-1, value less than low.value will be set to low.value.}

\item{high.value}{Feature value is standarded to 0-1, value greater than high.value will be set to high.value.}

\item{down.sample}{Number of cells to plot. Cells will be randomly downsampled for this number for plotting.}

\item{seed.use}{Random seed [10].}

\item{plot.nrow}{Number of rows in the plot [3].}

\item{plot.ncol}{Number of columns in the plot [3].}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{pdf.width}{the width of the graphics region in inches [7].}

\item{...}{Arguments passed to ggplot.}
}
\description{
Project Gene-body Accessibility to T-SNE plot
}
