% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotViz}
\alias{plotViz}
\title{Visulization}
\usage{
plotViz(obj, method, point.size, point.shape, point.alpha, text.add, text.size,
  text.color, text.halo.add, text.halo.color, text.halo.width, down.sample,
  pdf.file.name, pdf.width, pdf.height, ...)
}
\arguments{
\item{obj}{A snap object.}

\item{method}{Method used to plot c("tsne", "umap").}

\item{point.size}{Point size [1].}

\item{point.shape}{Point shape type [19].}

\item{point.alpha}{Point transparancy level [0.8].}

\item{text.add}{Whether to add cluster label text at the centroid of each cluster [TRUE].}

\item{text.size}{Cluster label text size [5].}

\item{text.color}{Cluster label text color ["black"].}

\item{text.halo.add}{Add halo to cluster label text [TRUE].}

\item{text.halo.color}{Halo color ["white"].}

\item{text.halo.width}{Halo width [0.5].}

\item{down.sample}{Downsample the original cells to down.sample cells to ovoid large dataset [10,000].}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.width}{the width of the graphics region in inches [7].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{...}{Arguments passed to plot method.}
}
\description{
Visulization
}
