% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictGenePeakPair.R
\name{predictGenePeakPair}
\alias{predictGenePeakPair}
\title{Link Distal Regulatory Elements (peaks) to Putative Target Genes}
\usage{
predictGenePeakPair(obj, input.mat = c("pmat", "bmat"), gene.name = NULL,
  gene.loci = NULL, do.par = FALSE, num.cores = 1)
}
\arguments{
\item{obj}{A snap object.}

\item{input.mat}{Input matrix c("pmat", "bmat").}

\item{gene.name}{Name of the target gene. It must be a gene in the cell-by-gene matrix [NULL].}

\item{gene.loci}{A genomic range object that contains the flanking window of the target gene [NULL].}

\item{do.par}{A logical variable indicates if to run this in parallel using multiple processors [FALSE].}

\item{num.cores}{A numeric class that indicates the number of cores to use for calculation [1].}
}
\value{
Return A data.frame that contains the peaks and associating P-value.
}
\description{
Link Distal Regulatory Elements (peaks) to Putative Target Genes
}
