% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapRbind.R
\name{rBind}
\alias{rBind}
\title{Combine snap objects}
\usage{
rBind(obj1, ...)
}
\arguments{
\item{obj1}{a snap object}

\item{obj2}{a snap object}
}
\value{
a combined snap object
}
\description{
Takes two snap objects and combines them.
}
\examples{
file <- system.file("extdata", "Fang.3C1.snap", package = "SNAPATAC");
x1.sp <- readSnap(file, binSize=5000, metaData=TRUE);
file <- system.file("extdata", "Fang.3C2.snap", package = "SNAPATAC");
x2.sp <- readSnap(file, binSize=5000, metaData=TRUE);
rBind(x1.sp, x2.sp);
}
