% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{readMetaData}
\alias{readMetaData}
\title{Read meta data from a snap file}
\usage{
readMetaData(file)
}
\arguments{
\item{file}{character for the snap-format file name which the data are to be read from.}
}
\value{
A data frame contains barcodes and its attributes
}
\description{
Take a snap file as input and read the barcode session only.
}
