% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{rmBmatFromSnap}
\alias{rmBmatFromSnap}
\title{Remove cell-by-bin matrix}
\usage{
rmBmatFromSnap(obj)
}
\arguments{
\item{obj}{A snap object to remove cell-by-bin matrix.}
}
\value{
Return a snap object without cell-by-bin matrix
}
\description{
This function takes a snap object as input and removes the cell-by-bin 
matrix in the existing snap object. Report error when cell-by-bin matrix is empty.
}
\examples{
data(demo.sp)
rmBmatFromSnap(demo.sp)

}
