% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{rmGmatFromSnap}
\alias{rmGmatFromSnap}
\title{Remove cell-by-gene matrix}
\usage{
rmGmatFromSnap(obj)
}
\arguments{
\item{obj}{A snap object to remove cell-by-gene matrix.}
}
\value{
Return a snap object without cell-by-peak matrix
}
\description{
This function takes a snap object as input and removes the cell-by-gene 
matrix in the existing snap object. Report error when cell-by-gene matrix 
is empty.
}
\examples{
data(demo.sp)
rmPmatFromSnap(demo.sp)

}
