% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{rowCovs}
\alias{rowCovs}
\alias{rowCovs,snap-method}
\alias{colCovs,snap-method}
\alias{rowCovs,snap-method}
\title{rowCovs for snap object.}
\usage{
rowCovs(x, ...)

\S4method{rowCovs}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = FALSE)

\S4method{colCovs}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = FALSE)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns the number of non-zero elements for each row of the count matrix
}
\examples{
barcode_num <- 100
feature_num <- 500
m <- matrix(sample(0:1,barcode_num * feature_num, replace=TRUE),barcode_num,feature_num)
x.sp <- newSnap(m)
barcode_cov <- rowCovs(x.sp)
}
