% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{rowMeans}
\alias{rowMeans}
\alias{rowMeans,snap-method}
\title{rowMeans for snap objects}
\usage{
\S4method{rowMeans}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = FALSE)
}
\arguments{
\item{x}{snap; a snap object}
}
\description{
This function takes a snap object and returns the row means of its count matrix.
}
