% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{rowSums}
\alias{rowSums}
\alias{rowSums,snap-method}
\title{rowSums for snap objects}
\usage{
\S4method{rowSums}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = FALSE)
}
\arguments{
\item{x}{snap; a snap object}
}
\description{
This function takes a snap object and returns the row sums of its count matrix.
}
