% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runChromVAR.R
\name{runChromVAR}
\alias{runChromVAR}
\title{Estimate Motif Variability Using ChromVAR}
\usage{
runChromVAR(obj, input.mat, genome, min.count, species)
}
\arguments{
\item{obj}{A snap object.}

\item{input.mat}{Input matrix used for chromVAR analysis c("pmat", "bmat").}

\item{genome}{BSgenome object that contains the sequence for the corresponding genome (i.e. BSgenome.Hsapiens.UCSC.hg19).}

\item{min.count}{Min count for features. Features of count lower than min.count will be filtered [10].}

\item{species}{Species or jaspar code used by 'getJasparMotifs'to get the motif database ["Homo sapiens"].}
}
\value{
Return a matrix object that contains the cell-by-motif matrix
}
\description{
Estimate Motif Variability Using ChromVAR
}
