% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClusters.R
\name{runCluster}
\alias{runCluster}
\title{Find Clusters Using Louvain Algorithm}
\usage{
runCluster(obj, louvain.lib, resolution, path.to.snaptools, load.knn.from.file,
  edge.file, seed.use)
}
\arguments{
\item{obj}{A snap object}

\item{louvain.lib}{Louvain implementation method to use ["R-igraph", "python-louvain"].
"R-igraph" uses "cluster_louvain" to identify cluster. "python-louvain" uses "python-louvain" package to identify
clusters. "R-igraph" is signficantly faster but does not support different resolutions. "python-louvain" supports
multiple resolutions, but requires "snaptools" to be pre-installed.}

\item{resolution}{A numeric class that indicates the resolution for louvain clustering [1]. This is effective only when path.to.snaptools is set.}

\item{path.to.snaptools}{Path to snaptools excutable file [NULL]. Install snaptools by 'pip install snaptools'.}

\item{load.knn.from.file}{Whether to load knn grph from a file [FALSE].}

\item{edge.file}{Directory path to the file that contains knn graphs [NULL].}

\item{seed.use}{Random seed [10].}
}
\value{
Returns a snap obj with the cluster stored in obj@cluster
}
\description{
Using the constructed knn graph, we next applied community finding algorithm Louvain 
to identify the ‘communities’ in the resulting graph which represents groups of
cells sharing similar profiles, potentially originating from the same cell type.
}
