% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDM.R
\name{runDiffusionMaps}
\alias{runDiffusionMaps}
\title{Dimentionality Reduction by Diffusion Maps Algorithm}
\usage{
runDiffusionMaps(obj, input.mat = c("bmat", "pmat"), num.eigs = 20)
}
\arguments{
\item{obj}{A snap obj}

\item{input.mat}{Input matrix c("bmat", "pmat").}

\item{num.eigs}{Number of eigenvectors to be computed [20].}
}
\description{
This function takes a snap obj as input and runs diffusion maps 
for dimentionality reduction.
}
\details{
Diffusion Maps algorithm, a nonlinear dimensionality reduction technique 
that discovers low dimensional manifolds within high-dimensional datasets 
by performing harmonic analysis of a random walk constructed over the data 
to identify nonlinear collective variables containing the predominance of 
the variance in the data. We choose diffusion maps because it is highly robust 
to noise and perturbation, making it particuarly suited for analyzing 
sparse scATAC-seq dataset.
}
\examples{
data(demo.sp);
demo.sp = makeBinary(demo.sp);
demo.sp = runDiffusionMaps(
obj=demo.sp, 
input.mat="bmat", 
num.eigs=20
);
}
