% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDimReduct.R
\name{runDimReduct}
\alias{runDimReduct}
\title{Linear Dimentionality Reduction}
\usage{
runDimReduct(obj, pc.num, input.mat, method, center, scale, seed.use, maxit,
  ...)
}
\arguments{
\item{obj}{A snap obj.}

\item{pc.num}{An integer number of dimetions to return [50].}

\item{input.mat}{A character value indicates what input matrix to use - could be c("nmat", "jmat", "bmat", "pmat") [nmat].}

\item{method}{A character value indicates what linear dimentionality reduction method to use - could be c("svd", "pca.whiten", "pca") [svd].}

\item{center}{A logical value indicating whether the variables should be centered to be zero [TRUE].}

\item{scale}{A logical value indicating whether the variables should be scaled to be standarized [FALSE].}

\item{seed.use}{A numeric integer value interpreted as an integer [10].}

\item{maxit}{Maximum number of iterations [1000].}

\item{...}{Arguments passed to irlba fuction.}
}
\description{
This function takes a snap obj as input and performs computation of linear dimentionality reduction method using irlba
}
\examples{
data(demo.sp);
demo.sp = runDimReduct(obj=demo.sp, pc.num=10, input.mat="jmat");

}
