% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMACS.R
\name{runMACS}
\alias{runMACS}
\title{Call Peaks Using MACS2}
\usage{
runMACS(obj, output.prefix, path.to.snaptools, path.to.macs, gsize,
  buffer.size = 500,
  macs.options = "--nomodel --shift 37 --ext 73 --qval 1e-2 -B --SPMR --call-summits",
  num.cores = 1, tmp.folder = NULL, keep.minimal = TRUE)
}
\arguments{
\item{obj}{A snap object.}

\item{output.prefix}{Prefix of output file which will be used to generate output file names.}

\item{path.to.snaptools}{Path to snaptools excutable file.}

\item{path.to.macs}{Path to macs2 excutable file.}

\item{gsize}{effective genome size. 'hs' for human, 'mm' for mouse, 'ce' for C. elegans, 'dm' for fruitfly (default: None)}

\item{buffer.size}{Buffer size for incrementally increasing internal array size to store reads alignment information. In
most cases, you don't have to change this parameter. However, if there are very high coverage dataset that each barcode has
more than 10000 fragments, it's recommended to specify a smaller buffer size in order to decrease memory usage (but it will 
take longer time to read snap files) [1000].}

\item{macs.options}{String indicate options you would like passed to macs2. strongly do not recommand to change unless you 
know what you are doing. the default is '--nomodel --shift 37 --ext 73 --qval 1e-2 -B --SPMR --call-summits'.}

\item{num.cores}{Number of cores to use for omputing [1].}

\item{tmp.folder}{Directory to store temporary files. If not given, snaptools will automatically generate a temporary location
to store temporary files.}

\item{keep.minimal}{Keep minimal version of output [TRUE].}
}
\value{
Return a data.frame object that contains the peak information
}
\description{
Identify peaks using selected cells. Fragments belonging to a subset of cells 
are extracted and used to identify peaks using MACS2. This function requires
"MACS2" and "snaptools" preinstalled and excutable.
}
