% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMagic.R
\name{runMagic}
\alias{runMagic}
\title{Affinity Graph-based Smoothing}
\usage{
runMagic(obj, input.mat = c("gmat", "pmat", "bmat", "mmat"), step.size = 3)
}
\arguments{
\item{obj}{A snap obj}

\item{input.mat}{Input matrix c("bmat", "pmat", "gmat", "mmat").}

\item{step.size}{Number of steps for diffusion [3].}
}
\description{
This function takes a snap object as input with bmat/pmat/gmat/mmat slot 
and run magic - an affinity graph-based method - to smooth the signal.
}
