% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCA.R
\name{runPCA}
\alias{runPCA}
\title{Linear Dimentionality Reduction}
\usage{
runPCA(object, ...)
}
\arguments{
\item{object}{A snap object.}

\item{pc.num}{An integer number of dimetions to return [50].}

\item{input.mat}{A character value indicates what input matrix to use - could be c("nmat", "jmat", "bmat", "pmat") [nmat];}

\item{method}{A character value indicates what linear dimentionality reduction method to use - could be c("svd", "pca.whiten", "pca") [svd]}

\item{center}{A logical value indicating whether the variables should be centered to be zero [TRUE]}

\item{scale}{A logical value indicating whether the variables should be scaled to be standarized [FALSE]}

\item{weight.by.sd}{A logical value indicating whether to scale the vector using standard deviation [FALSE];}

\item{seed.use}{a numeric integer value interpreted as an integer [10];}
}
\description{
This function takes a snap object as input and performs computation of linear dimentionality reduction method using irlba
}
