% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSnapAddPmat.R
\name{runSnapAddPmat}
\alias{runSnapAddPmat}
\title{Add cell-by-peak matrix to a snap file}
\usage{
runSnapAddPmat(obj, peak, path.to.snaptools, tmp.folder, num.cores, buffer.size)
}
\arguments{
\item{obj}{A snap obj.}

\item{peak}{A GenomicRanges object that contains peak coordinates.}

\item{path.to.snaptools}{Path to snaptools excutable file.}

\item{tmp.folder}{Directory to store temporary files generated by this program.}

\item{num.cores}{Number of CPUs to use.}

\item{buffer.size}{Buffer size for incrementally increasing internal array size to store reads alignment information. In
most cases, you don't have to change this parameter. However, if there are very high coverage dataset that each barcode has
more than 10000 fragments, it's recommended to specify a smaller buffer size in order to decrease memory usage (but it will take longer time to read snap files) [1000].}
}
\description{
This function takes a peak list and a snap object as input 
and add cell-by-peak matrix into the existing snap file.
}
