% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSVD.R
\name{runSVD}
\alias{runSVD}
\title{Principle Component Anaysis (PCA)}
\usage{
runSVD(x, ...)
}
\arguments{
\item{object}{A snap object.}

\item{pc.num}{An integer number of principal component vectors to return [50].}

\item{input.mat}{A character value indicates what input matrix to use could be c("nmat", "jmat", "cmat", "bmat") [nmat];}

\item{pc.features}{A numeric index indicates what variables to use [NULL];}

\item{center}{a logical value indicating whether the variables should be centered to be zero [TRUE]}

\item{scale}{a logical value indicating whether the variables should be scaled to be standarized [FALSE]}

\item{seed.use}{a numeric integer value interpreted as an integer [10];}
}
\description{
This function takes a snap object as input and performs computation of a truncated principal components analysis using irlba
}
