% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSNE.R
\name{runViz}
\alias{runViz}
\title{Dimentionality Reduction for Visualization}
\usage{
runViz(object, ...)
}
\arguments{
\item{object}{A snap object.}

\item{dims}{integer; Output dimensionality (default: 2)}

\item{pca_dims}{integer vector; dimentions of PCA used for tsne or umap}

\item{method}{character; A character variable indicates what t-sne method to use ("Rtsne", "fast_tsne", "umap").}

\item{fast_tsne_path}{character; A character variable indicates where the excutable fast_tsne located. Required when method="fast_tsne"}

\item{Y_init}{matrix; matrix; Initial locations of the objects. If NULL, random initialization will be used (default: NULL). Note that when using this, the initial stage with exaggerated perplexity values and a larger momentum term will be skipped.}

\item{seed.use}{numeric; random seed}

\item{tsne_perplexity}{numeric; perplexity for runing tsne}

\item{tsne_theta}{numeric; theta for tsne}

\item{max_tsne_iter}{numeric; max iteration for running tsne}
}
\value{
Returns a snap object with the tsne or umap
}
\description{
Wrapper for the implementation of Barnes-Hut t-Distributed
Stochastic Neighbor Embedding and FFT-accelerated Interpolation-based 
t-SNE (FIt-SNE) t-SNE, and UMAP
}
