% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{class}
\name{snap-class}
\alias{snap-class}
\title{An S4 class snap to represent single-nucleus accessibility object.}
\description{
Class snap defines a snap object.
}
\section{Slots}{

\describe{
\item{\code{des}}{A character object describes the details of the experiment.}

\item{\code{barcode}}{A character vector contains cell barcodes as rows.}

\item{\code{metaData}}{A data.frame object contains meta data for barcodes.}

\item{\code{feature}}{A GRanges object contains genomic features (bins).}

\item{\code{peak}}{A GRanges object contains genomic features (peaks).}

\item{\code{bmat}}{A Matrix object contains cellxbin count matrix.}

\item{\code{pmat}}{A Matrix object contains cellxpeak count matrix.}

\item{\code{gmat}}{A matrix object contains cellxgene count matrix.}

\item{\code{jmat}}{A jaccard object that contains jaccard index matrix.}

\item{\code{smat}}{A list object contains the linear dimentionality reduction result.}

\item{\code{graph}}{A kgraph object that contains the knn graph.}

\item{\code{tsne}}{A matrix object contains tsne coordinate.}

\item{\code{umap}}{A matrix object contains umap coordinate.}

\item{\code{cluster}}{A factor object contains cluster label.}
}}

