% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{snapListRbind}
\alias{snapListRbind}
\title{Combine snap objects}
\usage{
snapListRbind(snapList)
}
\arguments{
\item{snapList}{List of snap objects}
}
\value{
a combined snap object
}
\description{
We have checked this function behaviour with snapRbind.
They are identical.
}
