% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapToSeurat.R
\name{snapToSeurat}
\alias{snapToSeurat}
\title{Convert a snap object to a seurat object}
\usage{
snapToSeurat(obj, eigs.dims = 1:20, norm = TRUE, scale = TRUE)
}
\arguments{
\item{obj}{A snap object.}

\item{eigs.dims}{A vector of the dimensions to use.}

\item{norm}{A logical variable indicates whether to normalize the cell by gene matrix.}

\item{scale}{A logical variable indicagtes whether to scale the cell by gene matrix.}
}
\value{
Return a seurat object that contains single cell ATAC-seq data
}
\description{
Convert a snap object to a seurat object
}
