% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard-class.R
\docType{methods}
\name{[,jaccard,ANY,ANY,ANY-method}
\alias{[,jaccard,ANY,ANY,ANY-method}
\title{subsetting for jaccard objects}
\usage{
\S4method{[}{jaccard,ANY,ANY,ANY}(x, i, j, drop = "missing")
}
\arguments{
\item{x}{A jaccard object}

\item{i}{selected rows}

\item{j}{selected columns}

\item{drop}{drop unused levels}
}
\description{
This function takes a jaccard object and returns the subset of jaccard object.
}
\examples{
data(demo.sp);
demo.sub.sp = demo.sp[1:5,]
}
