% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgraph-class.R
\docType{methods}
\name{[,kgraph,ANY,ANY,ANY-method}
\alias{[,kgraph,ANY,ANY,ANY-method}
\title{subsetting for kgraph objects}
\usage{
\S4method{[}{kgraph,ANY,ANY,ANY}(x, i, j, drop = "missing")
}
\arguments{
\item{x}{A kgraph object}

\item{i}{selected rows}

\item{drop}{drop unused levels}
}
\description{
This function takes a kgraph object and returns the subset of kgraph object.
}
