% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\docType{methods}
\name{[,snap,ANY,ANY,ANY-method}
\alias{[,snap,ANY,ANY,ANY-method}
\title{subsetting for snap objects}
\usage{
\S4method{[}{snap,ANY,ANY,ANY}(x, i, j, mat = c("bmat", "pmat"),
  drop = "missing")
}
\arguments{
\item{x}{snap; a snap object}

\item{i}{integer; selected barcode index}

\item{j}{integer; selected feature index}

\item{mat}{character; indicates the slot to subsetting}

\item{drop}{character;}
}
\description{
This function takes a snap object and returns the subset of snap object.
}
