% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{Summary}
\alias{Summary}
\alias{Summary,snap-method}
\title{Summary for snap objects}
\usage{
\S4method{Summary}{snap}(x)
}
\arguments{
\item{x}{snap; a snap object}
}
\description{
This function takes a snap object and returns the quality control metrics summary.
}
