% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\docType{methods}
\name{summarySnap}
\alias{summarySnap}
\alias{summarySnap,snap-method}
\title{summarySnap for snap object.}
\usage{
summarySnap(obj)

\S4method{summarySnap}{snap}(obj)
}
\arguments{
\item{obj}{snap; a snap object}
}
\description{
This function takes a snap object and returns the summary statistics
}
\examples{
file.name = system.file("extdata", "demo.snap", package = "SnapATAC");
x.sp = createSnap(file.name, sample="demo");
summarySnap(x.sp);
}
