% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{addBmatToSnap}
\alias{addBmatToSnap}
\title{Add cell-by-bin matrix}
\usage{
addBmatToSnap(obj, file, bin.size)
}
\arguments{
\item{obj}{A snap object to add cell-by-bin matrix.}

\item{file}{A snap file.}

\item{bin.size}{Cell-by-bin matrix with bin size of bin.size will be added to snap object [5000].}
}
\value{
Return a snap object
}
\description{
This function takes a snap object and snap-format file as input and add the cell-by-bin 
matrix to the existing snap object.
}
