% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{addBmatToSnap}
\alias{addBmatToSnap}
\title{Add cell-by-bin matrix}
\usage{
addBmatToSnap(obj, bin.size, do.par, num.cores)
}
\arguments{
\item{obj}{A snap object to add cell-by-bin matrix.}

\item{bin.size}{Cell-by-bin matrix with bin size of bin.size 
will be added to snap object.}

\item{do.par}{A logical variable indicates whether use multiple processors [FALSE].}

\item{num.cores}{Number of processors to use [1].}
}
\value{
Return a snap object
}
\description{
This function takes a snap object as input and add the cell-by-bin 
matrix to the existing snap object.
}
\examples{
file.name = system.file("extdata", "demo.snap", package = "SnapATAC");
demo.sp = createSnap(file.name, sample="demo", do.par=FALSE);
showBinSizes(file.name);
demo.sp = addBmatToSnap(demo.sp, bin.size=100000, do.par=FALSE);

}
