% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{addPmatToSnap}
\alias{addPmatToSnap}
\title{Add cell-by-peak matrix}
\usage{
addPmatToSnap(obj, do.par, num.cores)
}
\arguments{
\item{obj}{A snap object.}

\item{do.par}{A logical varaible indicates whether to use multiple processors [FALSE].}

\item{num.cores}{Number of processors to use [1].}
}
\description{
This function takes a snap object as input and add the 
cell-by-peak matrix to the existing snap object.
}
\examples{
file.name = system.file("extdata", "demo.snap", package = "SnapATAC");
demo.sp = createSnap(file.name, sample="demo", do.par=FALSE);
showBinSizes(file.name);
demo.sp = addPmatToSnap(demo.sp, do.par=FALSE);

}
