% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{calCellGeneTable}
\alias{calCellGeneTable}
\title{Calculate cell-by-gene count matrix}
\usage{
calCellGeneTable(obj, gene, num.cores, mat)
}
\arguments{
\item{obj}{A snap object.}

\item{gene}{A GRanges object that contains the genomic intervals of genes.}

\item{num.cores}{Number of CPU processers for computation [1].}

\item{mat}{Matrix slot is used to calculatd the cell-by-gene matrix c("bmat", "pmat").}
}
\value{
Returns a Snap obj with the cell-by-gene matrix stored in obj@gmat
}
\description{
This function takes a snap obj and converts the cell-by-bin 
matrix into the cell-by-gene count matrix.
}
