% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-class.R
\docType{methods}
\name{colSums}
\alias{colSums}
\alias{colSums,snap-method}
\alias{colSums}
\alias{colSums,snap-method}
\title{colSums for snap objects}
\usage{
\S4method{colSums}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = FALSE)

FindOverlaps(x, y, ...)
}
\arguments{
\item{x}{snap; a snap object}

\item{x}{snap; a snap object}
}
\description{
This function takes a snap object and returns the column sums of its count matrix.

This function takes a snap object and returns the column sums of its count matrix.
}
