% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{createSnap}
\alias{createSnap}
\title{Create a snap object from a snap file}
\usage{
createSnap(file, metaData, description)
}
\arguments{
\item{file}{A snap-format file name.}

\item{metaData}{A logical value indicates whether read the meta data [TRUE].}

\item{description}{Description of the experiment [NULL].}
}
\value{
A snap object
}
\description{
This function takes a snap-format file as input and create
a snap object.
}
