% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{createSnap}
\alias{createSnap}
\title{Create a snap object from a snap file}
\usage{
createSnap(file, sample, description, num.cores)
}
\arguments{
\item{file}{Name of a snap-format file.}

\item{sample}{A short sample name (i.g. "MOS.rep1").}

\item{description}{Description of the experiment [NULL].}

\item{num.cores}{Number of processers to use [1].}
}
\value{
A snap object
}
\description{
This function takes a snap-format file as input and create
a snap object.
}
