% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnap.R
\name{createSnap}
\alias{createSnap}
\title{Create a Snap object from a snap file}
\usage{
createSnap(file, ...)
}
\arguments{
\item{file}{A character object for the snap-format file name which the data are to be read from.}

\item{metaData}{a logical value indicates weather to read the meta data [TRUE].}
}
\value{
A snap object
}
\description{
Import a snap file as a snap object.
}
\examples{
snap_file <- system.file("extdata", "Fang.3C1.snap", package = "SNAPATAC");
showBinSizes(snap_file);
x.sp <- createSnap(snap_file, metaData=TRUE);
}
