% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{createSnapFromBmat}
\alias{createSnapFromBmat}
\title{Create a snap object from cell-by-bin matrix}
\usage{
createSnapFromBmat(mat, barcodes, bins)
}
\arguments{
\item{mat}{A sparse matrix}

\item{barcodes}{Corresponding barcodes}

\item{bins}{A GenomicRanges object for the genomic coordinates of the bins}
}
\value{
Return a snap object
}
\description{
This function takes a cell-by-bin count matrix as input and returns a snap object.
}
