% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnap.R
\name{createSnapFromBmat}
\alias{createSnapFromBmat}
\title{Create a snap object from cell-by-bin matrix}
\usage{
createSnapFromBmat(mat, barcodes, bins)
}
\description{
Create a snap object from cell-by-bin matrix
}
