% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{createSnapFromGmat}
\alias{createSnapFromGmat}
\title{Create a snap object from cell-by-gene matrix}
\usage{
createSnapFromGmat(mat, barcodes, gene.names)
}
\arguments{
\item{mat}{A sparse matrix}

\item{barcodes}{An array of characters for barcodes}

\item{gene.names}{An array of characters for gene names}
}
\description{
This function takes a cell-by-gene count matrix as input and returns a snap object.
}
\examples{
gmat = Matrix(sample.int(100, 100), nrow=10, ncol=10, sparse=TRUE);
barcodes = paste("barcode", 1:10, sep=".");
genes = paste("gene", 1:10, sep=".");
createSnapFromGmat(gmat, barcodes, gene.names=genes);
}
