% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnap.R
\name{createSnapFromPmat}
\alias{createSnapFromPmat}
\title{Create a snap object from cell-by-peak matrix}
\usage{
createSnapFromPmat(mat, barcodes, peaks)
}
\description{
Create a snap object from cell-by-peak matrix
}
