% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{createSnapFromPmat}
\alias{createSnapFromPmat}
\title{Create a snap object from cell-by-peak matrix}
\usage{
createSnapFromPmat(mat, barcodes, peaks)
}
\arguments{
\item{mat}{A sparse matrix}

\item{barcodes}{Corresponding barcodes}

\item{peaks}{A GRanges object for the genomic coordinates of peaks}
}
\description{
This function takes a cell-by-peak count matrix as input and returns a snap object.
}
