% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBins.R
\name{filterBins}
\alias{filterBins}
\title{Feature Filteration}
\usage{
filterBins(object, ...)
}
\arguments{
\item{object}{snap object}

\item{low.threshold}{Low cutoffs for the parameters (default is -2)}

\item{high.threshold}{High cutoffs for the parameters (default is 2)}

\item{mat}{input matrix}
}
\value{
Returns a snap object containing only the relevant subset of bins
}
\description{
This function takes a snap object as input and filter feature based on given cutoff.
}
