% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{filterBins}
\alias{filterBins}
\title{Feature filtration}
\usage{
filterBins(obj, low.threshold, high.threshold, mat)
}
\arguments{
\item{obj}{A snap obj}

\item{low.threshold}{Low cutoffs for the parameters (default is -2)}

\item{high.threshold}{High cutoffs for the parameters (default is 2)}

\item{mat}{Matrix to filter c("bmat", "pmat")}
}
\value{
Returns a snap obj
}
\description{
This function takes a snap obj as input and perform feature selection in the following manner:
1) calculate coverage of each genomic bin/feature;
2) log scale the coverage by log10(count + 1);
3) the log-nromal distribution is then converted into zscore; 
4) bins with zscore beyond [low.threshold, high.threshold] are filtered;
}
\examples{
data(demo.sp);
filterBins(
obj=demo.sp, 
low.threshold=-2,
high.threshold=2
);

}
