% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDARs.R
\name{findDAR}
\alias{findDAR}
\title{Identifiy Statistically Singficant Differential Accessible Regions (DARs)}
\usage{
findDAR(object, ...)
}
\description{
Identifiy Statistically Singficant Differential Accessible Regions (DARs)
}
