% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\docType{methods}
\name{is.snap}
\alias{is.snap}
\alias{is.snap,snap-method}
\title{Check snap object}
\usage{
is.snap(obj)

\S4method{is.snap}{snap}(obj)
}
\arguments{
\item{obj}{A snap object}
}
\description{
This function takes any object as input and check if it is a snap object
}
