% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{isSnapFile}
\alias{isSnapFile}
\title{Check a snap-format file}
\usage{
isSnapFile(file)
}
\arguments{
\item{file}{A file name.}
}
\value{
Return a logical variable indicates whether file is a snap file.
}
\description{
This function takes a file name as input and check if the file is a snap-formated file
}
