% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\name{newSnap}
\alias{newSnap}
\title{Create an empty snap object}
\usage{
newSnap()
}
\value{
An empty snap object
}
\description{
This function creates an empty snap object
}
\examples{
x.sp = newSnap();

}
