% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcode.R
\name{plotBarcode}
\alias{plotBarcode}
\title{Plot Snap Barocde Stats}
\usage{
plotBarcode(obj)
}
\arguments{
\item{obj}{a snap object}
}
\description{
Plot the basic barcode statistics
}
\examples{
snap_file <- system.file("extdata", "Fang.3C1.snap", package = "nucleus");
showBinSizes(snap_file);
x.sp <- readSnap(snap_file, binSize=5000, metaData=TRUE);
plotBarcode(x.sp);
}
