% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotDimReductPW}
\alias{plotDimReductPW}
\title{Pairwise plot for Dimentionality Reduction Result}
\usage{
plotDimReductPW(obj, pca.dims, point.size, point.color, point.shape,
  point.alpha, down.sample, pdf.file.name, pdf.height, pdf.width)
}
\arguments{
\item{obj}{A snap object}

\item{pca.dims}{PC dimetions to plot [1:30]}

\item{point.size}{Point size [2].}

\item{point.color}{Point color ["grey"].}

\item{point.shape}{Point shape type [19].}

\item{point.alpha}{Point transparancy level [1].}

\item{down.sample}{Number of cells to plot. down.sample cells will be randomly downsampled for plot if there are more cells.}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{pdf.width}{the width of the graphics region in inches [7].}
}
\description{
Pairwise plot for Dimentionality Reduction Result
}
\examples{
data(demo.sp);
demo.sp = makeBinary(demo.sp, mat="bmat");
demo.sp = runJaccard(
obj=demo.sp, 
tmp.folder=tempdir(), 
mat="bmat"
);
demo.sp = runNormJaccard(
obj=demo.sp, 
tmp.folder=tempdir()
);
demo.sp = runDimReduct(
obj=demo.sp, 
pc.num=10, 
input.mat="jmat"
);
plotDimReductPW(demo.sp, pca.dims=1:10);

}
