% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotFeatureSingle}
\alias{plotFeatureSingle}
\title{plotFeatureSingle}
\usage{
plotFeatureSingle(obj, feature.value, method, point.size, point.shape,
  down.sample, pdf.file.name, pdf.width, pdf.height, quantiles, ...)
}
\arguments{
\item{obj}{A snap object.}

\item{feature.value}{Feature enrichment value for each cell. Value will be normalized betweeen 0 and 1.}

\item{method}{Visulization method c("tsne", "umap").}

\item{point.size}{Point size [1].}

\item{point.shape}{Point shape type [19].}

\item{down.sample}{Downsample the original cells to down.sample cells to ovoid large dataset [10,000].}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.width}{Width of the graphics region in inches [7].}

\item{pdf.height}{Height of the graphics region in inches [7].}

\item{quantiles}{Feature value outside this range will be removed [c(0.01, 0.99)]}

\item{...}{Arguments passed to plot method.}
}
\description{
plotFeatureSingle
}
\examples{
data(demo.sp);
x = rexp(nrow(demo.sp));
x = (x-min(x))/(max(x)-min(x));
PlotFeatureSingle(
	obj=demo.sp, 
	feature.value=x,
	method="tsne", 
	point.size=1, 
	point.shape=19, 
	down.sample=10000, 
	pdf.file.name=NULL, 
	pdf.width=7, 
	pdf.height=7
)

}
