% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotViz}
\alias{plotViz}
\title{Visulization}
\usage{
plotViz(obj, method, point.size, point.shape, point.alpha, point.color,
  text.add, text.size, text.color, text.halo.add, text.halo.color,
  text.halo.width, legend.add, legend.pos, legend.text.size, legend.text.color,
  down.sample, pdf.file.name, pdf.width, pdf.height, ...)
}
\arguments{
\item{obj}{A snap object.}

\item{method}{Visulization method c("tsne", "umap").}

\item{point.size}{Point size [1].}

\item{point.shape}{Point shape type [19].}

\item{point.alpha}{Point transparancy level [0.8].}

\item{point.color}{Color of point. Two options, points color by 
cell cluster label or sample ID c("cluster", "sample").}

\item{text.add}{Whether to add cluster label text at the centroid of each cluster [TRUE].}

\item{text.size}{Cluster label text size [5].}

\item{text.color}{Cluster label text color ["black"].}

\item{text.halo.add}{Add halo to cluster label text [TRUE].}

\item{text.halo.color}{Halo color ["white"].}

\item{text.halo.width}{Halo width [0.5].}

\item{legend.add}{Add a legend to the plot [FALSE].}

\item{legend.pos}{Position of the legend c("bottomleft", "bottom", "left", "topleft", "top", "topright", "right", "center").}

\item{legend.text.size}{Size of the text in the legend [1].}

\item{legend.text.color}{Color of the text in the legend ["black"].}

\item{down.sample}{Downsample the original cells to down.sample cells to ovoid large dataset [10,000].}

\item{pdf.file.name}{pdf file name to save the plot [NULL].}

\item{pdf.width}{the width of the graphics region in inches [7].}

\item{pdf.height}{the height of the graphics region in inches [7].}

\item{...}{Arguments passed to plot method.}
}
\description{
Visulization
}
\examples{
data(demo.sp);
plotViz(
	obj=demo.sp, 
	method="tsne", 
	point.color="sample", 
	text.add=FALSE,
legend.add=TRUE
);

}
