% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnap.R
\name{rBind}
\alias{rBind}
\title{Combine snap objects}
\usage{
rBind(obj1, obj2)
}
\arguments{
\item{obj1}{a snap object}

\item{obj2}{a snap object}
}
\value{
a combined snap object
}
\description{
Takes two snap objects and combines them.
}
