% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\docType{methods}
\name{rowMeans}
\alias{rowMeans}
\alias{rowMeans,snap-method}
\title{rowMeans for snap objects}
\usage{
\S4method{rowMeans}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = TRUE)
}
\arguments{
\item{x}{A snap object}

\item{mat}{A charater object indicates what matrix slot to use for calculation}

\item{na.rm}{A logical variable indicates wether to remove NA in the matrix}
}
\description{
This function takes a snap object and returns the row means of its count matrix.
}
\examples{
data(demo.sp);
a = rowMeans(demo.sp, mat="bmat");
b = rowMeans(demo.sp, mat="pmat");
d = rowMeans(demo.sp, mat="gmat");

}
