% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-utilities.R
\docType{methods}
\name{rowSums}
\alias{rowSums}
\alias{rowSums,snap-method}
\title{rowSums for snap objects}
\usage{
\S4method{rowSums}{snap}(x, mat = c("bmat", "pmat", "gmat"), na.rm = TRUE)
}
\arguments{
\item{x}{A snap object}

\item{mat}{A charater object indicates what matrix slot to use}

\item{na.rm}{A logical variable indicates wether to remove NA in the matrix}
}
\description{
This function takes a snap object and returns the row sums of its count matrix.
}
