% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClusters.R
\name{runCluster}
\alias{runCluster}
\title{Find Clusters Using Louvain Algorithm}
\usage{
runCluster(obj, louvain.lib, resolution, path.to.snaptools, seed.use)
}
\arguments{
\item{obj}{A snap object}

\item{louvain.lib}{Louvain implementation method to use ["R-igraph", "python-louvain"].
"R-igraph" uses "cluster_louvain" implemented by igraph package in R. "python-louvain" uses louvain method implemented in "python-louvain" package. 
"R-igraph" is signficantly faster but does not support different resolutions. "python-louvain" supports
multiple resolutions, but requires "snaptools" and runs significantly slower.}

\item{resolution}{A numeric value that indicates the resolution for louvain clustering [1]. This is effective only when path.to.snaptools is set.}

\item{path.to.snaptools}{Path to snaptools excutable file [NULL]. Install snaptools by 'pip install snaptools'.}

\item{seed.use}{Random seed [10].}
}
\value{
Returns a snap obj with the cluster stored in obj@cluster
}
\description{
Using the constructed knn graph returned by function runKNN, we next applied community finding algorithm Louvain 
to identify the ‘communities’ in the resulting graph which represents groups of
cells sharing similar accessibility profiles.
}
