% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{runCluster}
\alias{runCluster}
\title{Find Clusters Using Louvain Method}
\usage{
runCluster(object, ...)
}
\arguments{
\item{object}{A snap object.}

\item{k}{A numeric object indicates the size for KNN graphs [15].}

\item{pc.sel}{A numeric vector indicates which PCs are used for creating KNN graph [NULL].}

\item{func}{A character object indicates what louvain function to use "python" or "R" [python] .}

\item{method}{A character object indicates what cluster methods to use "louvain" or "jaccard.louvain" [louvain].}

\item{resolution}{A numeric vector between 0 and 1 indicates what resolution to use for louvain [0.5].}

\item{data_path}{A character variable indicates the path for storing graph data [NULL].}

\item{result_path}{A character variable indicates the path for storing clustering result [NULL].}

\item{community_path}{A character variable indicates the path to community_louvain program [bin/community_louvain].}
}
\value{
Returns a Snap object with the cluster stored in object@cluster
}
\description{
This function takes a snap object with "smat" attributes and identify cluster using graph-based clustering method
}
