% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHarmony.R
\name{runHarmony}
\alias{runHarmony}
\title{Perform batch effect using harmony}
\usage{
runHarmony(obj, eigs.dim, meta_data, vars_use, ...)
}
\arguments{
\item{obj}{A snap object}

\item{eigs.dim}{A vector of the components to use in construction of the KNN graph.}

\item{meta_data}{Either (1) Dataframe with variables to integrate or (2) vector with labels.}

\item{vars_use}{If meta_data is dataframe, this defined which variable(s) to remove (character vector).}

\item{...}{Paramters passed to Harmony}
}
\description{
Integration of single cell ATAC-seq data for batch correction using Harmony algorithm
}
