% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSnapAddPmat.R
\name{runSnapAddPmat}
\alias{runSnapAddPmat}
\title{Add cell-by-peak matrix to a snap file}
\usage{
runSnapAddPmat(file, peak, path.to.snaptools, buffer.size, tmp.folder)
}
\arguments{
\item{file}{A snap file.}

\item{peak}{A GenomicRanges object that contains peak coordinates.}

\item{path.to.snaptools}{Path to snaptools excutable file.}

\item{buffer.size}{Buffer size for incrementally increasing internal array size to store reads alignment information. In
most cases, you don't have to change this parameter. However, if there are very high coverage dataset that each barcode has
more than 10000 fragments, it's recommended to specify a smaller buffer size in order to decrease memory usage (but it will take longer time to read snap files) [1000].}

\item{tmp.folder}{Directory to store temporary files. If not given, snaptools will automatically generate a temporary location to store temporary files.}
}
\description{
This function takes a peak list and a snap file and add cell-by-peak matrix 
into the existing snap file.
}
