% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runViz.R
\name{runViz}
\alias{runViz}
\title{Dimentionality Reduction for Visualization}
\usage{
runViz(obj, dims, pca.dims, weight.by.sd, method, fast_tsne_path, Y.init,
  seed.use, num.cores, ...)
}
\arguments{
\item{obj}{A snap object.}

\item{dims}{integer; Output dimensionality (default: 2)}

\item{pca.dims}{integer vector; principal components used for running tsne or umap}

\item{weight.by.sd}{Weight the cell embeddings by the sd of each PC}

\item{method}{character; A character variable indicates what t-sne method to use ("Rtsne", "fast_tsne", "umap").}

\item{fast_tsne_path}{character; A character variable indicates the path of the excutable FIt-sen located. Required only when method="fast_tsne"}

\item{Y.init}{matrix; Initial locations of the objs. If NULL, random initialization will be used (default: NULL). Note that when using this, the initial stage with exaggerated perplexity values and a larger momentum term will be skipped.}

\item{seed.use}{number; Random seeds to use.}

\item{num.cores}{number; Number of CPU used for computing.}

\item{...}{Arguments passed to Rtsne, umap or FIt-tsne}
}
\value{
Returns a snap obj with the visulization
}
\description{
Wrapper for the implementation of Barnes-Hut t-Distributed
Stochastic Neighbor Embedding and FFT-accelerated Interpolation-based 
t-SNE (FIt-SNE) t-SNE, and UMAP
}
